requirements.txt
streamlit>=1.40.0
google-genai>=1.0.0
ollama>=0.4.0
python-dotenv>=1.0.0
